feat: move edit button next to the baytype selection
